+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 
 - Add ``data_sensitivity`` field to contract
 - Add ``get_sensitivity_rationales`` to retrieve sensitivity rationales
+- Add ``patch_dataset`` method
 
 2.1.0 - 2026-04-29
 ------------------